+++ royalflush/_behaviour/premiofl/train.py
@@ -23,6 +23,9 @@ class TrainAndApplyConsensusState(State):
             )
         self.agent.algorithm_logger.restart_chrono()
         self.agent.current_round += 1
+        self.agent.nn_convergence_logger.current_round = self.agent.current_round
+        self.agent.nn_train_logger.current_round = self.agent.current_round
+        self.agent.nn_inference_logger.current_round = self.agent.current_round
         if self.agent.are_max_iterations_reached():
             self.agent.logger.info(
                 f"[{self.agent.current_round - 1}] Stopping agent because max rounds "
@@ -38,20 +41,14 @@ class TrainAndApplyConsensusState(State):
                 # Train the model
                 self.agent.logger.debug(f"[{self.agent.current_round}] Starting training...")
                 metrics_train = self.agent.model_manager.train(
-                    train_logger=self.agent.nn_train_logger.log_train_epoch,
-                    weight_logger=self.agent.nn_convergence_logger.log_weight,
-                    agent_jid=self.agent.jid,
-                    current_round=self.agent.current_round,
-                    weight_id=-1,
+                    train_logger=self.agent.nn_train_logger,
+                    weight_logger=self.agent.nn_convergence_logger,
                 )
-
                 metrics_validation = self.agent.model_manager.inference()
                 metrics_test = self.agent.model_manager.test_inference()
-                self.log_model_results(
-                    trains=metrics_train,
-                    validation=metrics_validation,
-                    test=metrics_test,
-                )
+                self.agent.nn_inference_logger.log(metrics_validation=metrics_validation, metrics_test=metrics_test)
+
+                self.log_train_results(trains=metrics_train)
 
                 self.set_next_state("communication")
 
@@ -59,7 +56,7 @@ class TrainAndApplyConsensusState(State):
             self.agent.logger.exception(e)
             traceback.print_exc()
 
-    def log_model_results(self, trains: list[ModelMetrics], validation: ModelMetrics, test: ModelMetrics) -> None:
+    def log_train_results(self, trains: list[ModelMetrics]) -> None:
         if trains:
             start_t = trains[0].start_time_z
             end_t = trains[-1].end_time_z
@@ -71,17 +68,4 @@ class TrainAndApplyConsensusState(State):
                     f"[{self.agent.current_round}] Train completed in "
                     + f"{train_time.total_seconds():.2f} seconds with mean accuracy {mean_accuracy:.6f} and mean"
                     + f" loss {mean_loss:.6f} iterating {len(trains)} epochs."
-                )
-                self.agent.nn_inference_logger.log(
-                    current_round=self.agent.current_round,
-                    agent=self.agent.jid,
-                    seconds=train_time.total_seconds(),
-                    epochs=len(trains),
-                    mean_training_accuracy=mean_accuracy,
-                    mean_training_loss=mean_loss,
-                    validation_accuracy=validation.accuracy,
-                    validation_loss=validation.loss,
-                    test_accuracy=test.accuracy,
-                    test_loss=test.loss,
-                    timestamp=start_t,
                 )